refactor(RHICOMPL-1104): split GreySmallText
